+++ pkg/utils/fastthrift/fast_thrift_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fast_thrift
+package fastthrift
 
 import (
 	"testing"
